requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
spacy==3.7.6
sentence-transformers==3.0.1
scikit-learn==1.5.2
httpx==0.27.2
python-dateutil==2.9.0
aiohttp==3.10.5
pydantic==2.8.2
